Code/requirements.txt
asttokens==2.4.1
bm3d==4.0.3
bm4d==4.2.5
comm==0.2.2
contourpy==1.3.0
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
debugpy==1.8.7
decorator==5.1.1
executing==2.1.0
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
imageio==2.36.0
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.10.0.84
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
PyWavelets==1.7.0
pyzmq==26.2.0
scikit-image==0.24.0
scipy==1.14.1
setuptools==75.3.0
six==1.16.0
stack-data==0.6.3
sympy==1.13.1
tifffile==2024.9.20
tk==0.1.0
torch==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
triton==3.1.0
typing_extensions==4.12.2
wcwidth==0.2.13
